Relocate review rubric helpers
